+++ .gitignore
@@ -6,7 +6,6 @@ __pycache__/
 *secret.json
 *.h5
 requirements.txt
-tests/
 *.gz
 examples/imdb/aclImdb/
 examples/coco/coco/coco_data/
